backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.0
openai==1.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiohttp==3.9.1
pydantic-settings==2.1.0
